# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Iris To Do API",
    "version": "1.0.0",
    "description": "Documentación de Iris To Do API para prueba técnica"
  },
  "paths": {
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Obtiene una lista de tareas",
        "description": "Recupera todas las tareas almacenadas en el sistema.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tareas obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "ID de la tarea"
                      },
                      "title": {
                        "type": "string",
                        "description": "Título de la tarea"
                      },
                      "tag": {
                        "type": "string",
                        "description": "Etiqueta de la tarea"
                      },
                      "completed": {
                        "type": "boolean",
                        "description": "Estado de la tarea (completada o no)"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Crea una nueva tarea",
        "description": "Crea una nueva tarea proporcionando los detalles de la tarea en formato JSON.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Título de la tarea"
                  },
                  "tag": {
                    "type": "string",
                    "description": "Etiqueta de la tarea"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Estado de la tarea (completada o no)"
                  }
                },
                "required": [
                  "title",
                  "tag",
                  "completed"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarea creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID de la tarea creada"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título de la tarea"
                    },
                    "tag": {
                      "type": "string",
                      "description": "Etiqueta de la tarea"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Estado de la tarea (completada o no)"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de creación de la tarea"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de la última actualización de la tarea"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{taskId}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Obtiene una tarea por su ID",
        "description": "Recupera los detalles de una tarea específica utilizando su ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la tarea"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la tarea obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID de la tarea"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título de la tarea"
                    },
                    "tag": {
                      "type": "string",
                      "description": "Etiqueta de la tarea"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Estado de la tarea (completada o no)"
                    },
                    "required": [
                      "title",
                      "tag",
                      "completed"
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Actualiza una tarea existente",
        "description": "Actualiza los detalles de una tarea existente utilizando su ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la tarea"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Título de la tarea"
                  },
                  "tag": {
                    "type": "string",
                    "description": "Etiqueta de la tarea"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Estado de la tarea (completada o no)"
                  }
                },
                "required": [
                  "title",
                  "tag",
                  "completed"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID de la tarea"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título de la tarea"
                    },
                    "tag": {
                      "type": "string",
                      "description": "Etiqueta de la tarea"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Estado de la tarea (completada o no)"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de creación de la tarea"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de la última actualización de la tarea"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Elimina una tarea por su ID",
        "description": "Elimina una tarea específica utilizando su ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la tarea"
          }
        ],
        "responses": {
          "200": {
            "description": "Tarea eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de confirmación"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tasks/complete/{taskId}": {
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Marca una tarea como completada",
        "description": "Actualiza el estado de una tarea a completada utilizando su ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la tarea"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "completed": {
                    "type": "boolean",
                    "description": "Estado de la tarea (completada o no)"
                  }
                },
                "required": [
                  "completed"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarea actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID de la tarea"
                    },
                    "title": {
                      "type": "string",
                      "description": "Título de la tarea"
                    },
                    "tag": {
                      "type": "string",
                      "description": "Etiqueta de la tarea"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Estado de la tarea (completada o no)"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de creación de la tarea"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha y hora de la última actualización de la tarea"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Tarea no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error en el servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica a un usuario y devuelve un token de acceso",
        "description": "Verifica las credenciales del usuario y, si son correctas, devuelve un token de acceso JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Nombre de usuario"
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña del usuario"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario autenticado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "Token de acceso JWT"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra un nuevo usuario y devuelve un token de acceso",
        "description": "Crea un nuevo usuario en el sistema y devuelve un token de acceso JWT.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Nombre de usuario"
                  },
                  "password": {
                    "type": "string",
                    "description": "Contraseña del usuario"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string",
                      "description": "Token de acceso JWT"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error en la solicitud",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensaje de error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}